backend/pyproject.toml
[tool.poetry]
name = "backend"
package-mode = false
version = "0.1.0"
description = "Backend for the Kagami research project."
authors = ["james.brandt <bran1400@umn.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = ">=0.115.13,<0.116.0"
uvicorn = { extras = ["standard"], version = ">=0.34.3,<0.35.0" }
python-multipart = ">=0.0.20,<0.0.21"
python-dotenv = ">=1.1.0,<2.0.0"
pydantic = ">=2.11.7,<3.0.0"
pydantic-settings = ">=2.10.0,<3.0.0"
httpx = ">=0.28.1,<0.29.0"
nltk = ">=3.9.1,<4.0.0"
spacy = ">=3.8.7,<4.0.0"
textstat = ">=0.7.7,<0.8.0"
emoji = ">=2.14.1,<3.0.0"
empath = ">=0.89,<0.90"
torch = { version = ">=2.7.0" }
transformers = ">=4.52.4,<5.0.0"
openai = ">=1.90.0,<2.0.0"
google-api-python-client = ">=2.173.0,<3.0.0"
google-auth = ">=2.40.3,<3.0.0"
google-auth-oauthlib = ">=1.2.2,<2.0.0"
requests = ">=2.32.4,<3.0.0"
psutil = ">=7.0.0,<8.0.0"
sentence-transformers = ">=4.1.0,<5.0.0"
pillow = ">=11.2.1,<12.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."